+++ src/Parser/Statement/ifStatement.java
@@ -7,7 +7,7 @@ import java.util.Map;
 
 public record ifStatement(Expr ex, Statement then, Statement els) implements Statement {
     @Override
-    public void eval(Map<String, Integer> bindings) throws EvalError {
+    public void eval(Map<String, Long> bindings) throws EvalError {
         if(ex.eval(bindings) > 0){
             then.eval(bindings);
         }else {
